some stuff

+++ admin/adminIncludes/viewAllposts.php
@@ -123,7 +123,7 @@
                     echo "<td>{$post_date}</td>";
                     echo "<td><a href='../post.php?p_id=$post_id'>View post</a></td>";
                     echo "<td><a href='Posts.php?source=editPost&p_id={$post_id}'>Edit</a></td>";
-                    echo "<td><a href='Posts.php?delete={$post_id}'>Delete</a></td>";
+                    echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this?'); \" href='Posts.php?delete={$post_id}'>Delete</a></td>";
                     echo "</tr>";
